HRMS_CORE_BACKEND/requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.7
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
mysql-connector-python==9.1.0
packaging==24.1
pip-tools==7.4.1
psycopg2==2.9.9
PyJWT==2.9.0
pyproject_hooks==1.2.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
